city_brain_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
requests==2.31.0
openai==1.3.5
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.28.1
pytest==8.4.2